requirements.txt
python-chess
opencv-python
numpy
pyautogui
mss
